src/requirements.txt
joblib==1.4.2
matplotlib==3.10.0
numpy==2.2.0
pandas==2.2.3
python-constraint==1.4.0
scikit_learn==1.6.0
seaborn==0.13.2
